create struct_with_invaraints test
