+++ source/rust_verify_test/tests/user_defined_type_invariants.rs
@@ -2272,27 +2272,6 @@ test_verify_one_file! {
     } => Err(err) => assert_vir_error_msg(err, "expected generics to match")
 }
 
-test_verify_one_file! {
-    #[test] struct_with_invariants_const_usage verus_code! {
-        use vstd::prelude::*;
-        use vstd::atomic_ghost::AtomicUsize;
-
-        const ONE: usize = 1;
-
-        struct_with_invariants!{
-            pub struct S {
-                x: AtomicUsize<_,(),_>,
-            }
-
-            closed spec fn wf(self) -> bool {
-                invariant on x is (v:usize,g:()) {
-                    v == ONE
-                }
-            }
-        }
-    } => Ok(())
-}
-
 test_verify_one_file_with_options! {
     #[test] mut_ref_not_supported [] => verus_code! {
         struct A {
@@ -2721,29 +2700,4 @@ test_verify_one_file_with_options! {
             set_to(Tracked(&mut x.i.0), 15); // FAILS
         }
     } => Err(err) => assert_fails_type_invariant_error(err, 2)
-}
-
-test_verify_one_file! {
-    #[test] struct_with_invariants_assoc_type_prefix verus_code! {
-        use core::marker::PhantomData;
-        use vstd::atomic_ghost::AtomicPtr;
-        use vstd::prelude::*;
-
-        pub trait HasTarget {
-            type Target;
-        }
-
-        struct_with_invariants!{
-            pub struct S<A: HasTarget>  {
-                a: AtomicPtr<<A as HasTarget>::Target,_,(),_>,
-                _phantom: PhantomData<*const A::Target>,
-            }
-
-            closed spec fn wf(self) -> bool {
-                invariant on a with (_phantom) is (v:*mut <A as HasTarget>::Target,g:()) {
-                    true
-                }
-            }
-        }
-    } => Ok(())
 }
